Variable size halos, split halo logic
Halo is a (too much) simple component and we cannot script it, i.e. we can only enable a Halo (by its Behaviour) but not customize size and color.

As we do not want to assign a Halo to each Interactable object (for maintenabiliy), we need to figure out a trick. It's not fully satisfying but there it is :
- Create a prefab with an empty GameObject with a Light component. The Light will not display (culling mask = 0), but has the drawHalo tick box checked. Internally, Unity will create a Halo with appropriate size.
- Now, in scripts, we can adjust the range property of the Light component, and Unity will re-create internally (with some obscure magic) the Halo with custom size and color.
- At each raycasting on a Interactable object, we compute a suitable size for the Halo, affect it to the Light, and set the instance of the empty GameObject as a child of the Interactable object.

That's not the most elegant way to manage variable size halos, but the best I could come up with. The PlayerInteraction script should hold a reference to the Prefab and call HaloController.Init(<instance of prefab>), the rest is managed by the later script.