Reorganized project / scene hierarchy
Project and scene hierarchy reorganized to better deal with assets.

In project window, all imported assets are into Objects/Imported Assets. In this folder, each object or semantic family of objects has a folder. Inside, there is :
- A folder Textures ;
- A folder Materials ;
- A folder Models ;
- A folder Prefabs.

If there is only one object, Models and Prefabs are replaced with <name>Obj and <name>Textured files. If there is no texture or material... well, no Textures or Materials folder.

The non-imported Materials are into Materials/Custom Materials.
The imported but used in multiple objets materials are into Materials/Standard Materials.

The objects we created are into Custom Prefabs.