Add piano keys animations
Piano keys extends InteractionBase and propose an animation when an interaction of type InteractionType.Use occurs.

The same animation has been used for every single key (great) because all keys have transform.position set to (0,0,0) (not great because halo position would be incorrect). This is due to some obscure dark magic with imported models and we will deal with it.