Take interaction, tutorial, bug fixes, VR/PC switch easier
New take interaction : middle button pushed on PC. Stop holding to drop. Just creates a fixed joint bewteen takable object and player. This mechanism is not suited for PC but will be ok for VR (given the wand is moving and we need to touch the objet to take it !). Takable objects must have a rigidbody.

Tutorial finished : new font / color / don't travel walls. Sound when player advances. Game presentation, halo explanation, observe example (book), use example (lamp), take example (key).

No more switch for opening attic trapdoor. Now take the key and simultaneously use the lock on the wall.

Added a trigger which closes the trapdoor when passing it downwards.

Bug fixes : attic lamp flikers ok and material's emission is managed.

Changes in event/interaction system to reflect the take implentation : in particular generic methods EndObserve() and EndTake() to differentiate. Eg when you do not raycast anymore an object, EndObserve(), but do not EndTake(). EndTake() only if an object was selected and key up.

No more ugly HasKey property : InteractionManager manages authorizations to use objects (coupled to game scenario) in CanUse, eg if the source's rigidbody has a FixedJoint and the gameobject fixed is a key, we authorize to open the corresponding door, we refuse otherwise.

Few changes in scaling, positions, etc.